fastback-end/requirements.txt
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
asttokens==2.1.0
audioread==3.0.0
backcall==0.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorednoise==2.1.0
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
executing==1.2.0
fastai==2.7.10
fastapi==0.85.2
fastaudio==0.1.4
fastcore==1.5.27
fastdownload==0.0.7
fastprogress==1.0.3
fonttools==4.38.0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.8.0
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
numba==0.56.3
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pooch==1.6.0
preshed==3.0.8
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
resampy==0.4.2
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
soundfile==0.11.0
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
stack-data==0.6.0
starlette==0.20.4
thinc==8.1.5
threadpoolctl==3.1.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
traitlets==5.5.0
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
wasabi==0.10.1
wcwidth==0.2.5
zipp==3.10.0
